Refactor: 캘린더를 FullCalendar 컴포넌트로 분리
- 캘린더 폭 CSS 통일 위한 class명 부여(container)
